requirements.txt
flask==2.3.3
twilio==8.10.0
PyJWT==2.8.0
python-dotenv==1.0.0
